+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Reply to button in incoming logs so you can easily reply to a contact
+
 ### Changed
 - Dropdowns now do full-text search (searching by surname will now work)